refactor: multiple home page scaffolding
