+++ apps/picsa-apps/extension-app/src/app/pages/home/components/welcome/welcome.component.html
@@ -1,0 +1,3 @@
+<div style="margin: auto; max-width: 500px">
+  <h2>{{ 'Choose language' | translate }}</h2>
+</div>
