+++ apps/picsa-apps/extension-app/src/app/pages/home/components/welcome/welcome.component.scss
@@ -1,0 +1,3 @@
+h2 {
+  text-align: center;
+}
