+++ apps/picsa-apps/extension-app/src/app/pages/home/components/welcome/welcome.component.ts
@@ -1,0 +1,12 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+import { PicsaTranslateModule } from '@picsa/shared/modules';
+
+@Component({
+  selector: 'picsa-welcome',
+  standalone: true,
+  imports: [CommonModule, PicsaTranslateModule],
+  templateUrl: './welcome.component.html',
+  styleUrl: './welcome.component.scss',
+})
+export class PicsaWelcomeComponent {}
