--- apps/picsa-apps/extension-app/src/app/pages/home/home.page.scss
@@ -1,117 +1,0 @@
-.page {
-  background: var(--color-primary);
-}
-.page-content {
-  overflow: hidden;
-  padding: 0;
-}
-.background-image {
-  position: relative;
-}
-// note - could use media breakpoints for different images
-.background-image::before {
-  background-image: url('../../../assets/images/cover.jpg');
-  background-size: cover;
-  background-position: center;
-  content: '';
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: -2;
-  opacity: 0.4;
-}
-.background-image::after {
-  background-color: var(--color-primary);
-  content: '';
-  display: block;
-  position: absolute;
-  top: 0px;
-  left: 0px;
-  width: 100%;
-  height: 100%;
-  z-index: -1;
-  opacity: 0.4;
-}
-.link-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(128px, 1fr));
-  grid-auto-rows: max-content;
-  gap: 8px;
-  flex: 1;
-  padding: 1rem;
-  margin-bottom: 3rem;
-  overflow-y: auto;
-  overflow-x: hidden;
-  max-height: calc(100vh - 136px);
-}
-
-.grid-tile {
-  position: relative;
-  background: rgba(0, 0, 0, 0.65);
-  color: white;
-  cursor: pointer;
-  font-size: 18px;
-  padding: 16px 8px 8px 8px;
-}
-.grid-tile[data-dev-only='true'] {
-  background: rgba(0, 0, 0, 0.35);
-}
-.grid-tile:hover {
-  transform: scale(1.02);
-  transition: all 0.3s ease-out;
-}
-
-.tile-badge {
-  position: absolute;
-  top: -8px;
-  right: -6px;
-  background: var(--color-secondary);
-  border-radius: 50%;
-  height: 24px;
-  width: 24px;
-  line-height: 24px;
-  padding: 4px;
-  font-size: 16px;
-  text-align: center;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  font-weight: bold;
-  letter-spacing: 2px;
-}
-
-footer {
-  display: flex;
-  justify-content: space-between;
-  padding: 5px;
-  position: absolute;
-  bottom: 0;
-  width: calc(100% - 10px);
-  left: 0;
-  background: var(--color-primary);
-  align-items: center;
-}
-
-.nav-icon {
-  width: auto;
-  height: 70px;
-}
-.nav-text {
-  min-height: 3rem;
-  max-height: 4rem;
-  overflow: auto;
-  text-overflow: clip;
-  word-wrap: break-word;
-}
-.share-button {
-  padding: 0;
-}
-.version-info,
-.privacy-policy {
-  color: rgba(255, 255, 255, 0.7);
-}
-.privacy-policy {
-  text-decoration: none;
-}
